requirements.txt
# المكتبات الأساسية للبوت
python-telegram-bot==20.7
aiohttp==3.9.1
aiofiles==23.2.1

# مكتبات الـ Scraping
beautifulsoup4==4.12.2
cloudscraper==1.2.71
lxml==4.9.3
requests==2.31.0

# مكتبات المعالجة
Pillow==10.1.0
python-dotenv==1.0.0

# مكتبات المساعدة
asyncio==3.4.3
urllib3==2.1.0
